del dia 4 -09

+++ database/migrations/2025_05_31_152650_create_transfer_products_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->unsignedInteger('company_id')->default(1);
             $table->unsignedInteger('warehouse_id');
-            $table->unsignedTinyInteger('local_id');
+            $table->unsignedInteger('local_id');
             $table->unsignedInteger('user_id');
             $table->unsignedInteger('product_id');
             $table->decimal('output', 8,2)->nullable();
